+++ contracts/Dinngo.sol
@@ -35,6 +35,9 @@ contract Dinngo is SerializableOrder, SerializableWithdrawal, SerializableMigrat
     mapping (address => uint256) public tokenRanks;
     mapping (address => uint256) public lockTimes;
 
+    address public walletOwner;
+    address public DGOToken;
+
     event AddUser(uint256 userID, address indexed user);
     event AddToken(uint256 tokenID, address indexed token);
     event Deposit(address token, address indexed user, uint256 amount, uint256 balance);
@@ -217,17 +220,60 @@ contract Dinngo is SerializableOrder, SerializableWithdrawal, SerializableMigrat
     }
 
     /**
+     * @notice The extract function for fee in ether. Can only be triggered by
+     * the dinngo wallet owner.
+     * @param amount The amount to be withdrawn
+     */
+    function extractFee(uint256 amount) external {
+        require(amount > 0);
+        require(msg.sender == walletOwner);
+        balances[address(0)][address(0)] = balances[address(0)][address(0)].sub(amount);
+        msg.sender.transfer(amount);
+    }
+
+    /**
+     * @notice The extract function for fee in token. Can only be triggered by
+     * the dinngo wallet owner.
+     * @param token The token contract address to be withdrawn
+     * @param amount The amount to be withdrawn
+     */
+    function extractTokenFee(address token, uint256 amount) external {
+        require(amount > 0);
+        require(msg.sender ==  walletOwner);
+        require(token != address(0));
+        balances[token][address(0)] = balances[token][address(0)].sub(amount);
+        IERC20(token).safeTransfer(msg.sender, amount);
+    }
+
+    /**
+     * @notice The function to get the balance from fee account.
+     * @param token The token of the balance to be queried
+     */
+    function getWalletBalance(address token) external view returns (uint256) {
+        return balances[token][address(0)];
+    }
+
+    /**
+     * @notice The function to change the owner of fee wallet.
+     * @param newOwner The new wallet owner to be assigned
+     */
+    function changeWalletOwner(address newOwner) external {
+        require(newOwner != address(0));
+        require(newOwner != walletOwner);
+        walletOwner = newOwner;
+    }
+
+    /**
      * @notice The withdraw function that can only be triggered by owner.
      * Event Withdraw will be emitted after execution.
      * @param withdrawal The serialized withdrawal data
      */
     function withdrawByAdmin(bytes calldata withdrawal) external {
         address payable user = userID_Address[_getWithdrawalUserID(withdrawal)];
-        address wallet = userID_Address[0];
         address token = tokenID_Address[_getWithdrawalTokenID(withdrawal)];
         uint256 amount = _getWithdrawalAmount(withdrawal);
         uint256 amountFee = _getWithdrawalFee(withdrawal);
-        address tokenFee = _isWithdrawalFeeETH(withdrawal)? address(0) : tokenID_Address[1];
+        address tokenFee = _isWithdrawalFeeETH(withdrawal)? address(0) : DGOToken;
         uint256 balance = balances[token][user].sub(amount);
         require(_isValidUser(user));
         _verifySig(
@@ -242,8 +288,8 @@ contract Dinngo is SerializableOrder, SerializableWithdrawal, SerializableMigrat
         } else {
             balances[tokenFee][user] = balances[tokenFee][user].sub(amountFee);
         }
-        balances[tokenFee][wallet] =
-            balances[tokenFee][wallet].add(amountFee);
+        balances[tokenFee][address(0)] =
+            balances[tokenFee][address(0)].add(amountFee);
         balances[token][user] = balance;
         emit Withdraw(token, user, amount, balance, tokenFee, amountFee);
         if (token == address(0)) {
@@ -363,7 +409,6 @@ contract Dinngo is SerializableOrder, SerializableWithdrawal, SerializableMigrat
         require(amountBase != 0);
         // Get parameters
         address user = userID_Address[_getOrderUserID(order)];
-        address wallet = userID_Address[0];
         bytes32 hash = _getOrderHash(order);
         address tokenQuote = tokenID_Address[_getOrderTokenIDQuote(order)];
         address tokenBase = tokenID_Address[_getOrderTokenIDBase(order)];
@@ -382,24 +427,24 @@ contract Dinngo is SerializableOrder, SerializableWithdrawal, SerializableMigrat
             if (_isOrderFeeMain(order)) {
                 tokenFee = tokenBase;
                 balances[tokenBase][user] = balances[tokenBase][user].add(amountBase).sub(amountFee);
-                balances[tokenBase][wallet] = balances[tokenBase][wallet].add(amountFee);
+                balances[tokenBase][address(0)] = balances[tokenBase][address(0)].add(amountFee);
             } else {
-                tokenFee = tokenID_Address[1];
+                tokenFee = DGOToken;
                 balances[tokenBase][user] = balances[tokenBase][user].add(amountBase);
                 balances[tokenFee][user] = balances[tokenFee][user].sub(amountFee);
-                balances[tokenFee][wallet] = balances[tokenFee][wallet].add(amountFee);
+                balances[tokenFee][address(0)] = balances[tokenFee][address(0)].add(amountFee);
             }
         } else {
             balances[tokenBase][user] = balances[tokenBase][user].sub(amountBase);
             if (_isOrderFeeMain(order)) {
                 tokenFee = tokenQuote;
                 balances[tokenQuote][user] = balances[tokenQuote][user].add(amountQuote).sub(amountFee);
-                balances[tokenQuote][wallet] = balances[tokenQuote][wallet].add(amountFee);
+                balances[tokenQuote][address(0)] = balances[tokenQuote][address(0)].add(amountFee);
             } else {
-                tokenFee = tokenID_Address[1];
+                tokenFee = DGOToken;
                 balances[tokenQuote][user] = balances[tokenQuote][user].add(amountQuote);
                 balances[tokenFee][user] = balances[tokenFee][user].sub(amountFee);
-                balances[tokenFee][wallet] = balances[tokenFee][wallet].add(amountFee);
+                balances[tokenFee][address(0)] = balances[tokenFee][address(0)].add(amountFee);
             }
         }
         // Order fill