+++ contracts/DinngoProxy.sol
@@ -25,22 +25,28 @@ contract DinngoProxy is Ownable, Administrable, TimelockUpgradableProxy {
     mapping (address => uint256) public tokenRanks;
     mapping (address => uint256) public lockTimes;
 
+    address public walletOwner;
+    address public DGOToken;
+
     /**
      * @dev User ID 0 is the management wallet.
      * Token ID 0 is ETH (address 0). Token ID 1 is DGO.
-     * @param dinngoWallet The main address of dinngo
-     * @param dinngoToken The contract address of DGO
+     * @param _walletOwner The fee wallet owner
+     * @param _dinngoToken The contract address of DGO
+     * @param _impl The implementation contract address
      */
     constructor(
-        address payable dinngoWallet,
-        address dinngoToken,
-        address impl
-    ) Proxy(impl) public {
+        address payable _walletOwner,
+        address _dinngoToken,
+        address _impl
+    ) Proxy(_impl) public {
         processTime = 90 days;
-        userID_Address[0] = dinngoWallet;
-        userRanks[dinngoWallet] = 255;
+        walletOwner = _walletOwner;
         tokenID_Address[0] = address(0);
-        tokenID_Address[1] = dinngoToken;
+        tokenRanks[address(0)] = 1;
+        tokenID_Address[1] = _dinngoToken;
+        tokenRanks[_dinngoToken] = 1;
+        DGOToken = _dinngoToken;
     }
 
     /**
@@ -193,6 +199,54 @@ contract DinngoProxy is Ownable, Administrable, TimelockUpgradableProxy {
     function withdrawToken(address token, uint256 amount) external {
         (bool ok,) = _implementation().delegatecall(
             abi.encodeWithSignature("withdrawToken(address,uint256)", token, amount)
+        );
+        require(ok);
+    }
+
+    /**
+     * @notice The function to extract the fee from the fee account. This function can
+     * only be triggered by the income wallet owner.
+     * @param amount The amount to be extracted
+     */
+    function extractFee(uint256 amount) external {
+        (bool ok,) = _implementation().delegatecall(
+            abi.encodeWithSignature("extractFee(uint256)", amount)
+        );
+        require(ok);
+    }
+
+    /**
+     * @notice The function to extract the fee from the fee account. This function can
+     * only be triggered by the income wallet owner.
+     * @param token The token to be extracted
+     * @param amount The amount to be extracted
+     */
+    function extractTokenFee(address token, uint256 amount) external {
+        (bool ok,) = _implementation().delegatecall(
+            abi.encodeWithSignature("extractTokenFee(address,uint256)", token, amount)
+        );
+        require(ok);
+    }
+
+    /**
+     * @notice The function to get the balance from fee account.
+     * @param token The token of the balance to be queried
+     */
+    function getWalletBalance(address token) external returns (uint256 balance) {
+        (bool ok, bytes memory ret) = _implementation().delegatecall(
+            abi.encodeWithSignature("getWalletBalance(address)", token)
+        );
+        require(ok);
+        balance = abi.decode(ret, (uint256));
+    }
+
+    /**
+     * @notice The function to change the owner of fee wallet.
+     * @param newOwner The new wallet owner to be assigned
+     */
+    function changeWalletOwner(address newOwner) external onlyOwner {
+        (bool ok,) = _implementation().delegatecall(
+            abi.encodeWithSignature("changeWalletOwner(address)", newOwner)
         );
         require(ok);
     }